requirements.txt
gspread==6.1.2
google-auth==2.29.0
yfinance==0.2.38
pandas==2.2.3
numpy>=2.2.3,<3.0
requests==2.32.3
scipy==1.14.1
tqdm==4.66.4
python-dotenv==1.0.1
